refactor(core): extract BlockEditFeature from KeyDownController
