+++ packages/core/src/features/block-editing/index.ts
@@ -1,0 +1,2 @@
+export {BlockEditFeature} from './BlockEditFeature'
+export {getCaretRawPosInBlock, getDomRawPos, getDomRawPosInMark, setCaretAtRawPos} from './rawPosition'
