+++ packages/core/src/features/input/KeyDownController.ts
@@ -1,21 +1,11 @@
-import {childAt, htmlChildren, isHtmlElement, isTextNode, nextText} from '../../shared/checkers'
+import {isHtmlElement} from '../../shared/checkers'
 import type {NodeProxy} from '../../shared/classes'
 import {KEYBOARD} from '../../shared/constants'
-import {Caret} from '../caret'
 import {captureMarkupPaste, consumeMarkupPaste, getBoundaryOffset} from '../clipboard'
-import {addDragRow, getMergeDragRowJoinPos, mergeDragRows, canMergeRows} from '../drag/operations'
-import {isTextTokenSpan} from '../editable'
-import {createRowContent} from '../editing'
 import {deleteMark} from '../editing/utils/deleteMark'
 import {shiftFocusNext, shiftFocusPrev} from '../navigation'
-import type {MarkToken, Token} from '../parsing'
 import {isFullSelection, selectAllText} from '../selection'
 import type {Store} from '../store/Store'
-
-function isTextLikeRow(token: Token): boolean {
-	if (token.type === 'text') return true
-	return token.descriptor.hasSlot && token.descriptor.segments.length === 1
-}
 
 export class KeyDownController {
 	#keydownHandler?: (e: KeyboardEvent) => void
@@ -31,16 +21,14 @@ export class KeyDownController {
 		if (!container) return
 
 		this.#keydownHandler = e => {
-			if (e.key === KEYBOARD.LEFT) {
-				if (!this.#handleBlockArrowLeftRight(e, 'left')) shiftFocusPrev(this.store, e)
-			} else if (e.key === KEYBOARD.RIGHT) {
-				if (!this.#handleBlockArrowLeftRight(e, 'right')) shiftFocusNext(this.store, e)
-			} else if (e.key === KEYBOARD.UP || e.key === KEYBOARD.DOWN) {
-				this.#handleArrowUpDown(e)
+			if (!this.store.state.drag.get()) {
+				if (e.key === KEYBOARD.LEFT) {
+					shiftFocusPrev(this.store, e)
+				} else if (e.key === KEYBOARD.RIGHT) {
+					shiftFocusNext(this.store, e)
+				}
+				this.#handleDelete(e)
 			}
-
-			this.#handleDelete(e)
-			this.#handleEnter(e)
 			selectAllText(this.store, e)
 		}
 
@@ -74,354 +62,54 @@ export class KeyDownController {
 
 	#handleDelete(event: KeyboardEvent) {
 		const {focus} = this.store.nodes
-		const isDragMode = !!this.store.state.drag.get()
 
-		// Mark/span deletion only applies in non-drag mode.
-		// In drag mode the focus target is a block div, not a span/mark.
-		if (!isDragMode && (event.key === KEYBOARD.DELETE || event.key === KEYBOARD.BACKSPACE)) {
-			if (focus.isMark) {
-				if (focus.isEditable) {
-					if (event.key === KEYBOARD.BACKSPACE && !focus.isCaretAtBeginning) return
-					if (event.key === KEYBOARD.DELETE && !focus.isCaretAtEnd) return
-				}
-				event.preventDefault()
-				deleteMark('self', this.store)
-				return
-			}
+		if (event.key !== KEYBOARD.DELETE && event.key !== KEYBOARD.BACKSPACE) return
 
-			if (event.key === KEYBOARD.BACKSPACE) {
-				if (focus.isSpan && focus.isCaretAtBeginning && focus.prev.target) {
-					event.preventDefault()
-					deleteMark('prev', this.store)
-					return
-				}
+		if (focus.isMark) {
+			if (focus.isEditable) {
+				if (event.key === KEYBOARD.BACKSPACE && !focus.isCaretAtBeginning) return
+				if (event.key === KEYBOARD.DELETE && !focus.isCaretAtEnd) return
 			}
-
-			if (event.key === KEYBOARD.DELETE) {
-				if (focus.isSpan && focus.isCaretAtEnd && focus.next.target) {
-					event.preventDefault()
-					deleteMark('next', this.store)
-					return
-				}
-			}
-
-			if (focus.isSpan && focus.isEditable && window.getSelection()?.isCollapsed) {
-				const content = focus.content
-				const caret = focus.caret
-				if (event.key === KEYBOARD.BACKSPACE && caret > 0) {
-					event.preventDefault()
-					focus.content = content.slice(0, caret - 1) + content.slice(caret)
-					focus.caret = caret - 1
-					this.store.events.change.emit()
-					return
-				}
-				if (event.key === KEYBOARD.DELETE && caret >= 0 && caret < content.length) {
-					event.preventDefault()
-					focus.content = content.slice(0, caret) + content.slice(caret + 1)
-					focus.caret = caret
-					this.store.events.change.emit()
-					return
-				}
-			}
+			event.preventDefault()
+			deleteMark('self', this.store)
+			return
 		}
 
-		if (!isDragMode) return
-
-		const container = this.store.refs.container
-		if (!container) return
-
-		const blockDivs = htmlChildren(container)
-		const blockIndex = blockDivs.findIndex(
-			div => div === document.activeElement || div.contains(document.activeElement)
-		)
-		if (blockIndex === -1) return
-
-		const rows = this.store.state.tokens.get()
-		if (blockIndex >= rows.length) return
-
-		const token = rows[blockIndex]
-		const value = this.store.state.previousValue.get() ?? this.store.state.value.get() ?? ''
-		if (!this.store.state.onChange.get()) return
-
 		if (event.key === KEYBOARD.BACKSPACE) {
-			const blockDiv = blockDivs[blockIndex]
-			const caretAtStart = Caret.getCaretIndex(blockDiv) === 0
-
-			const blockText = 'content' in token ? token.content : ''
-			if (blockText === '') {
+			if (focus.isSpan && focus.isCaretAtBeginning && focus.prev.target) {
 				event.preventDefault()
-				const newValue =
-					rows.length <= 1
-						? ''
-						: (() => {
-								if (blockIndex >= rows.length - 1)
-									return value.slice(0, rows[blockIndex - 1].position.end)
-								return (
-									value.slice(0, rows[blockIndex].position.start) +
-									value.slice(rows[blockIndex + 1].position.start)
-								)
-							})()
-				this.store.state.innerValue.set(newValue)
-				queueMicrotask(() => {
-					const targetIndex = Math.max(0, blockIndex - 1)
-					const target = childAt(container, targetIndex)
-					if (target) {
-						target.focus()
-						Caret.setCaretToEnd(target)
-					}
-				})
-				return
-			}
-
-			if (caretAtStart && blockIndex > 0) {
-				const prevToken = rows[blockIndex - 1]
-				const currToken = rows[blockIndex]
-				if (canMergeRows(prevToken, currToken)) {
-					event.preventDefault()
-					const joinPos = getMergeDragRowJoinPos(rows, blockIndex)
-					const newValue = mergeDragRows(value, rows, blockIndex)
-					this.store.state.innerValue.set(newValue)
-					queueMicrotask(() => {
-						const target = childAt(container, blockIndex - 1)
-						if (target) {
-							target.focus()
-							const updatedRows = this.store.state.tokens.get()
-							const updatedToken = updatedRows[blockIndex - 1]
-							setCaretAtRawPos(target, updatedToken, joinPos)
-						}
-					})
-					return
-				}
-				event.preventDefault()
-				queueMicrotask(() => {
-					const target = blockDivs[blockIndex - 1]
-					target.focus()
-					if (prevToken.type !== 'mark') Caret.setCaretToEnd(target)
-				})
+				deleteMark('prev', this.store)
 				return
 			}
 		}
 
 		if (event.key === KEYBOARD.DELETE) {
-			const blockDiv = blockDivs[blockIndex]
-			const caretIndex = Caret.getCaretIndex(blockDiv)
-			const caretAtEnd = caretIndex === blockDiv.textContent.length
-			const caretAtStart = caretIndex === 0
-
-			if (caretAtStart && blockIndex > 0) {
-				const prevToken = rows[blockIndex - 1]
-				const currToken = rows[blockIndex]
-				if (canMergeRows(prevToken, currToken)) {
-					event.preventDefault()
-					const joinPos = getMergeDragRowJoinPos(rows, blockIndex)
-					const newValue = mergeDragRows(value, rows, blockIndex)
-					this.store.state.innerValue.set(newValue)
-					queueMicrotask(() => {
-						const target = childAt(container, blockIndex - 1)
-						if (target) {
-							target.focus()
-							const updatedRows = this.store.state.tokens.get()
-							const updatedToken = updatedRows[blockIndex - 1]
-							setCaretAtRawPos(target, updatedToken, joinPos)
-						}
-					})
-					return
-				}
+			if (focus.isSpan && focus.isCaretAtEnd && focus.next.target) {
 				event.preventDefault()
-				queueMicrotask(() => {
-					const target = blockDivs[blockIndex - 1]
-					target.focus()
-					if (prevToken.type !== 'mark') Caret.setCaretToEnd(target)
-				})
+				deleteMark('next', this.store)
 				return
 			}
+		}
 
-			if (caretAtEnd && blockIndex < rows.length - 1) {
-				const currToken = rows[blockIndex]
-				const nextToken = rows[blockIndex + 1]
-				if (canMergeRows(currToken, nextToken)) {
-					event.preventDefault()
-					const joinPos = getMergeDragRowJoinPos(rows, blockIndex + 1)
-					const newValue = mergeDragRows(value, rows, blockIndex + 1)
-					this.store.state.innerValue.set(newValue)
-					queueMicrotask(() => {
-						const target = childAt(container, blockIndex)
-						if (target) {
-							target.focus()
-							const updatedRows = this.store.state.tokens.get()
-							const updatedToken = updatedRows[blockIndex]
-							setCaretAtRawPos(target, updatedToken, joinPos)
-						}
-					})
-					return
-				}
+		if (focus.isSpan && focus.isEditable && window.getSelection()?.isCollapsed) {
+			const content = focus.content
+			const caret = focus.caret
+			if (event.key === KEYBOARD.BACKSPACE && caret > 0) {
 				event.preventDefault()
-				queueMicrotask(() => {
-					const target = blockDivs[blockIndex + 1]
-					target.focus()
-					Caret.trySetIndex(target, 0)
-				})
+				focus.content = content.slice(0, caret - 1) + content.slice(caret)
+				focus.caret = caret - 1
+				this.store.events.change.emit()
+				return
+			}
+			if (event.key === KEYBOARD.DELETE && caret >= 0 && caret < content.length) {
+				event.preventDefault()
+				focus.content = content.slice(0, caret) + content.slice(caret + 1)
+				focus.caret = caret
+				this.store.events.change.emit()
 				return
 			}
 		}
 	}
-
-	#handleEnter(event: KeyboardEvent) {
-		const isDragMode = !!this.store.state.drag.get()
-		if (!isDragMode) return
-		if (event.key !== KEYBOARD.ENTER) return
-		if (event.shiftKey) return
-
-		const container = this.store.refs.container
-		if (!container) return
-
-		const activeElement = document.activeElement
-		if (!isHtmlElement(activeElement) || !container.contains(activeElement)) return
-
-		event.preventDefault()
-
-		const blockDivs = htmlChildren(container)
-		let blockIndex = -1
-		for (let i = 0; i < blockDivs.length; i++) {
-			if (blockDivs[i] === activeElement || blockDivs[i].contains(activeElement)) {
-				blockIndex = i
-				break
-			}
-		}
-		if (blockIndex === -1) return
-
-		const rows = this.store.state.tokens.get()
-		const token = rows[blockIndex]
-		const blockDiv = blockDivs[blockIndex]
-		const value = this.store.state.previousValue.get() ?? this.store.state.value.get() ?? ''
-
-		if (!this.store.state.onChange.get()) return
-
-		const newRowContent = createRowContent(this.store.state.options.get())
-
-		if (!isTextLikeRow(token)) {
-			const newValue = addDragRow(value, rows, blockIndex, newRowContent)
-			this.store.state.innerValue.set(newValue)
-			queueMicrotask(() => {
-				const newBlockIndex = blockIndex + 1
-				if (newBlockIndex < container.children.length) {
-					const newBlockEl = childAt(container, newBlockIndex)
-					if (newBlockEl) {
-						newBlockEl.focus()
-						Caret.trySetIndex(newBlockEl, 0)
-					}
-				}
-			})
-			return
-		}
-
-		const absolutePos = getCaretRawPosInBlock(blockDiv, token)
-		const newValue = value.slice(0, absolutePos) + newRowContent + value.slice(absolutePos)
-		this.store.state.innerValue.set(newValue)
-
-		queueMicrotask(() => {
-			const newBlockIndex = blockIndex + 1
-			if (newBlockIndex < container.children.length) {
-				const newBlockEl = childAt(container, newBlockIndex)
-				if (newBlockEl) {
-					newBlockEl.focus()
-					Caret.trySetIndex(newBlockEl, 0)
-				}
-			}
-		})
-	}
-
-	#handleBlockArrowLeftRight(event: KeyboardEvent, direction: 'left' | 'right'): boolean {
-		if (!this.store.state.drag.get()) return false
-
-		const container = this.store.refs.container
-		if (!container) return false
-
-		const activeElement = document.activeElement
-		if (!isHtmlElement(activeElement) || !container.contains(activeElement)) return false
-
-		const blockDivs = htmlChildren(container)
-		const blockIndex = blockDivs.findIndex(div => div === activeElement || div.contains(activeElement))
-		if (blockIndex === -1) return false
-
-		const blockDiv = blockDivs[blockIndex]
-
-		if (direction === 'left') {
-			if (Caret.getCaretIndex(blockDiv) !== 0) return false
-			if (blockIndex === 0) return true
-			event.preventDefault()
-			const prevBlock = blockDivs[blockIndex - 1]
-			prevBlock.focus()
-			Caret.setCaretToEnd(prevBlock)
-			return true
-		}
-
-		const caretIndex = Caret.getCaretIndex(blockDiv)
-		const textLen = blockDiv.textContent.length
-		if (caretIndex !== textLen) return false
-		if (blockIndex >= blockDivs.length - 1) return true
-		event.preventDefault()
-		const nextBlock = blockDivs[blockIndex + 1]
-		nextBlock.focus()
-		Caret.trySetIndex(nextBlock, 0)
-		return true
-	}
-
-	#handleArrowUpDown(event: KeyboardEvent) {
-		if (!this.store.state.drag.get()) return
-
-		const container = this.store.refs.container
-		if (!container) return
-
-		const activeElement = document.activeElement
-		if (!isHtmlElement(activeElement) || !container.contains(activeElement)) return
-
-		const blockDivs = htmlChildren(container)
-		const blockIndex = blockDivs.findIndex(div => div === activeElement || div.contains(activeElement))
-		if (blockIndex === -1) return
-
-		const blockDiv = blockDivs[blockIndex]
-
-		if (event.key === KEYBOARD.UP) {
-			if (!Caret.isCaretOnFirstLine(blockDiv)) return
-			if (blockIndex === 0) return
-
-			event.preventDefault()
-			const caretRect = Caret.getCaretRect()
-			const caretX = caretRect?.left ?? blockDiv.getBoundingClientRect().left
-			const prevBlockDiv = blockDivs[blockIndex - 1]
-			prevBlockDiv.focus()
-			const prevRect = prevBlockDiv.getBoundingClientRect()
-			Caret.setAtX(prevBlockDiv, caretX, prevRect.bottom - 4)
-		} else if (event.key === KEYBOARD.DOWN) {
-			if (!Caret.isCaretOnLastLine(blockDiv)) return
-			if (blockIndex >= blockDivs.length - 1) return
-
-			event.preventDefault()
-			const caretRect = Caret.getCaretRect()
-			const caretX = caretRect?.left ?? blockDiv.getBoundingClientRect().left
-			const nextBlockDiv = blockDivs[blockIndex + 1]
-			nextBlockDiv.focus()
-			const nextRect = nextBlockDiv.getBoundingClientRect()
-			Caret.setAtX(nextBlockDiv, caretX, nextRect.top + 4)
-		}
-	}
-}
-
-/**
- * Computes the raw value offset (index into the full value string) corresponding
- * to the current caret position within `blockDiv`.
- *
- * Delegates to `getDomRawPos` using the current selection's focus node and offset.
- */
-function getCaretRawPosInBlock(blockDiv: HTMLElement, token: Token): number {
-	const selection = window.getSelection()
-	if (!selection?.rangeCount) return token.position.end
-
-	const {focusNode, focusOffset} = selection
-	if (!focusNode) return token.position.end
-
-	return getDomRawPos(focusNode, focusOffset, blockDiv, token)
 }
 
 export function handleBeforeInput(store: Store, event: InputEvent): void {
@@ -438,18 +126,11 @@ export function handleBeforeInput(store: Store, event: InputEvent): void {
 	}
 	if (selecting === 'all') store.state.selecting.set(undefined)
 
-	// In drag mode the focus target is a block div, not a text span.
-	// Block-level keys (Enter, Backspace, Delete) are handled by KeyDownController.
-	// Text insertions and in-block deletions need special handling to update state.
-	if (store.state.drag.get()) {
-		handleBlockBeforeInput(store, event)
-		return
-	}
+	if (store.state.drag.get()) return
 
 	const {focus} = store.nodes
 	if (!focus.target || !focus.isEditable) return
 
-	// Intercept markput paste before span-level handling to update raw value directly
 	if (
 		(event.inputType === 'insertFromPaste' || event.inputType === 'insertReplacementText') &&
 		handleMarkputSpanPaste(store, focus, event)
@@ -462,11 +143,6 @@ export function handleBeforeInput(store: Store, event: InputEvent): void {
 	}
 }
 
-/**
- * Handles paste with markput data in non-drag mode.
- * Updates the raw value directly (which triggers re-parsing) instead of
- * inserting markup text into a single span (which would cause recursive DOM updates).
- */
 function handleMarkputSpanPaste(store: Store, focus: NodeProxy, event: InputEvent): boolean {
 	const container = store.refs.container
 	if (!container) return false
@@ -498,9 +174,6 @@ function handleMarkputSpanPaste(store: Store, focus: NodeProxy, event: InputEven
 	const newValue = currentValue.slice(0, rawInsertPos) + markup + currentValue.slice(rawEndPos)
 	store.state.innerValue.set(newValue)
 
-	// Find which text token contains caretPos in the re-parsed token array.
-	// Use isNext+childIndex so FocusController navigates to childAt(childIndex+2)
-	// even after the anchor span is replaced by re-render.
 	const newTokens = store.state.tokens.get()
 	let targetIdx = newTokens.findIndex(
 		t => t.type === 'text' && caretPos >= t.position.start && caretPos <= t.position.end
@@ -621,296 +294,4 @@ export function replaceAllContentWith(store: Store, newContent: string): void {
 			firstChild.focus()
 		}
 	})
-}
-
-/**
- * Handles `beforeinput` events when the editor is in drag mode.
- * Intercepts text insertion and in-block deletion to update the raw value via
- * `store.state.innerValue.set`, since `applySpanInput` is designed for span-level editing only.
- * Block-level operations (Enter, Backspace/Delete at boundaries) are handled by
- * `KeyDownController` via `keydown` events.
- */
-function handleBlockBeforeInput(store: Store, event: InputEvent): void {
-	const container = store.refs.container
-	if (!container) return
-
-	const activeElement = document.activeElement
-	if (!isHtmlElement(activeElement) || !container.contains(activeElement)) return
-
-	const blockDivs = htmlChildren(container)
-	const blockIndex = blockDivs.findIndex(div => div === activeElement || div.contains(activeElement))
-	if (blockIndex === -1) return
-
-	const blockDiv = blockDivs[blockIndex]
-	const rows = store.state.tokens.get()
-	if (blockIndex >= rows.length) return
-
-	const token = rows[blockIndex]
-	const value = store.state.previousValue.get() ?? store.state.value.get() ?? ''
-
-	const focusAndSetCaret = (newRawPos: number) => {
-		queueMicrotask(() => {
-			const target = childAt(container, blockIndex)
-			if (!target) return
-			target.focus()
-			const updatedRows = store.state.tokens.get()
-			const updatedToken = updatedRows[blockIndex]
-			setCaretAtRawPos(target, updatedToken, newRawPos)
-		})
-	}
-
-	switch (event.inputType) {
-		case 'insertText': {
-			event.preventDefault()
-			const data = event.data ?? ''
-			const ranges = event.getTargetRanges()
-			let rawFrom: number
-			let rawTo: number
-			if (ranges.length > 0) {
-				const rawStart = getDomRawPos(ranges[0].startContainer, ranges[0].startOffset, blockDiv, token)
-				const rawEnd = getDomRawPos(ranges[0].endContainer, ranges[0].endOffset, blockDiv, token)
-				;[rawFrom, rawTo] = rawStart <= rawEnd ? [rawStart, rawEnd] : [rawEnd, rawStart]
-			} else {
-				rawFrom = rawTo = getCaretRawPosInBlock(blockDiv, token)
-			}
-			store.state.innerValue.set(value.slice(0, rawFrom) + data + value.slice(rawTo))
-			focusAndSetCaret(rawFrom + data.length)
-			break
-		}
-		case 'insertFromPaste':
-		case 'insertReplacementText': {
-			event.preventDefault()
-			const markup = store.refs.container ? consumeMarkupPaste(store.refs.container) : undefined
-			const pasteData = markup ?? event.dataTransfer?.getData('text/plain') ?? ''
-			const ranges = event.getTargetRanges()
-			let rawFrom: number
-			let rawTo: number
-			if (ranges.length > 0) {
-				const rawStart = getDomRawPos(ranges[0].startContainer, ranges[0].startOffset, blockDiv, token)
-				const rawEnd = getDomRawPos(ranges[0].endContainer, ranges[0].endOffset, blockDiv, token)
-				;[rawFrom, rawTo] = rawStart <= rawEnd ? [rawStart, rawEnd] : [rawEnd, rawStart]
-			} else {
-				rawFrom = rawTo = getCaretRawPosInBlock(blockDiv, token)
-			}
-			store.state.innerValue.set(value.slice(0, rawFrom) + pasteData + value.slice(rawTo))
-			focusAndSetCaret(rawFrom + pasteData.length)
-			break
-		}
-		case 'deleteContentBackward':
-		case 'deleteContentForward':
-		case 'deleteWordBackward':
-		case 'deleteWordForward':
-		case 'deleteSoftLineBackward':
-		case 'deleteSoftLineForward': {
-			const ranges = event.getTargetRanges()
-			if (!ranges.length) return
-			const rawStart = getDomRawPos(ranges[0].startContainer, ranges[0].startOffset, blockDiv, token)
-			const rawEnd = getDomRawPos(ranges[0].endContainer, ranges[0].endOffset, blockDiv, token)
-			const [rawFrom, rawTo] = rawStart <= rawEnd ? [rawStart, rawEnd] : [rawEnd, rawStart]
-			if (rawFrom === rawTo) return
-			event.preventDefault()
-			store.state.innerValue.set(value.slice(0, rawFrom) + value.slice(rawTo))
-			focusAndSetCaret(rawFrom)
-			break
-		}
-	}
-}
-
-/**
- * Recursively sets the caret inside a mark token's rendered DOM element.
- *
- * For simple marks (no children), the mark renders its nested content as a single
- * text node; we position the caret at `rawAbsolutePos - nested.start` within it.
- *
- * For complex marks with nested tokens (e.g. a heading that spans multiple inline
- * marks), we walk the mark element's childNodes in parallel with token children,
- * preferring later tokens at boundaries so a position equal to a mark's end
- * resolves into the following sibling text rather than the mark itself.
- *
- * Returns true when the caret was successfully placed, false when the position
- * could not be resolved (caller should fall back to end-of-block).
- */
-function setCaretInMarkAtRawPos(markElement: HTMLElement, markToken: MarkToken, rawAbsolutePos: number): boolean {
-	const sel = window.getSelection()
-	if (!sel) return false
-
-	let tokenIdx = 0
-	for (const childNode of Array.from(markElement.childNodes)) {
-		if (tokenIdx >= markToken.children.length) break
-		const tokenChild = markToken.children[tokenIdx]
-
-		if (isHtmlElement(childNode) && tokenChild.type === 'text') {
-			if (!isTextTokenSpan(childNode)) continue
-			if (rawAbsolutePos >= tokenChild.position.start && rawAbsolutePos <= tokenChild.position.end) {
-				const rawTextNode = childNode.firstChild
-				const textNode = isTextNode(rawTextNode) ? rawTextNode : null
-				const offset = rawAbsolutePos - tokenChild.position.start
-				if (textNode) {
-					const range = document.createRange()
-					range.setStart(textNode, Math.min(offset, textNode.length))
-					range.collapse(true)
-					sel.removeAllRanges()
-					sel.addRange(range)
-				} else {
-					const range = document.createRange()
-					range.setStart(childNode, 0)
-					range.collapse(true)
-					sel.removeAllRanges()
-					sel.addRange(range)
-				}
-				return true
-			}
-			tokenIdx++
-		} else if (isTextNode(childNode) && tokenChild.type === 'text') {
-			if (rawAbsolutePos >= tokenChild.position.start && rawAbsolutePos <= tokenChild.position.end) {
-				const offset = Math.min(rawAbsolutePos - tokenChild.position.start, childNode.length)
-				const range = document.createRange()
-				range.setStart(childNode, offset)
-				range.collapse(true)
-				sel.removeAllRanges()
-				sel.addRange(range)
-				return true
-			}
-			tokenIdx++
-		} else if (isHtmlElement(childNode) && tokenChild.type === 'mark') {
-			const nextChild = tokenIdx + 1 < markToken.children.length ? markToken.children[tokenIdx + 1] : null
-			const atBoundary =
-				rawAbsolutePos === tokenChild.position.end && nextChild?.position.start === rawAbsolutePos
-			if (
-				!atBoundary &&
-				rawAbsolutePos >= tokenChild.position.start &&
-				rawAbsolutePos <= tokenChild.position.end
-			) {
-				return setCaretInMarkAtRawPos(childNode, tokenChild, rawAbsolutePos)
-			}
-			tokenIdx++
-		}
-	}
-
-	return false
-}
-
-/**
- * Sets the caret at an absolute raw-value position within a drag row's DOM element.
- * For mark tokens, recursively searches nested content via setCaretInMarkAtRawPos.
- */
-function setCaretAtRawPos(blockDiv: HTMLElement, token: Token, rawAbsolutePos: number): void {
-	const sel = window.getSelection()
-	if (!sel) return
-
-	if (token.type === 'mark') {
-		if (setCaretInMarkAtRawPos(blockDiv, token, rawAbsolutePos)) return
-		Caret.setCaretToEnd(blockDiv)
-		return
-	}
-
-	// Text token: position caret directly in the text node
-	const offsetWithinToken = rawAbsolutePos - token.position.start
-	const walker = document.createTreeWalker(blockDiv, 4 /* SHOW_TEXT */)
-	const textNode = nextText(walker)
-	if (textNode) {
-		const charOffset = Math.min(offsetWithinToken, textNode.length)
-		const range = document.createRange()
-		range.setStart(textNode, charOffset)
-		range.collapse(true)
-		sel.removeAllRanges()
-		sel.addRange(range)
-		return
-	}
-
-	Caret.setCaretToEnd(blockDiv)
-}
-
-/**
- * Maps a DOM (node, offset) position to an absolute raw-value offset.
- * Walks up from `node` to find the direct child of `blockDiv`, then maps
- * to the corresponding token's raw position. For mark tokens with nested
- * content, recursively resolves the position within the mark's children.
- */
-function getDomRawPos(node: Node, offset: number, blockDiv: HTMLElement, token: Token): number {
-	if (node === blockDiv) {
-		const sel = window.getSelection()
-		if (sel?.focusNode && sel.focusNode !== blockDiv) {
-			return getDomRawPos(sel.focusNode, sel.focusOffset, blockDiv, token)
-		}
-		return token.position.end
-	}
-
-	// When the text node is a direct child of blockDiv (mark tokens without
-	// DraggableBlock wrapper), resolve position using the token directly.
-	if (node.nodeType === Node.TEXT_NODE && node.parentElement === blockDiv) {
-		if (token.type === 'mark') {
-			return getDomRawPosInMark(node, offset, blockDiv, token)
-		}
-		return token.position.start + Math.min(offset, token.content.length)
-	}
-
-	let child: Node | null = node.nodeType === Node.ELEMENT_NODE ? node : node.parentElement
-	while (child && child.parentElement !== blockDiv) {
-		child = child.parentElement
-	}
-	if (!child) return token.position.end
-
-	// For mark tokens (no DragMark wrapper), delegate to getDomRawPosInMark.
-	if (token.type === 'mark') {
-		return getDomRawPosInMark(node, offset, blockDiv, token)
-	}
-
-	// Text token
-	return token.position.start + Math.min(offset, token.content.length)
-}
-
-/**
- * Recursively maps a DOM (node, offset) position to an absolute raw-value offset
- * within a mark token's nested content. Handles marks that contain other marks
- * (e.g. h1 containing bold), correctly mapping cursor positions through the
- * nested DOM/token structure.
- *
- * Key rules:
- * - cursor at offset === 0 → raw position at mark start
- * - cursor at offset === full nested length → raw position at mark end (after closing delimiter)
- * - cursor in the middle → raw position within nested content
- */
-function getDomRawPosInMark(node: Node, offset: number, markElement: HTMLElement, markToken: MarkToken): number {
-	if (markToken.children.length === 0) {
-		if (offset === 0) return markToken.position.start
-		const nestedLen = markToken.slot?.content.length ?? markToken.value.length
-		if (nestedLen > 0 && offset >= nestedLen) {
-			if (markToken.content.endsWith('\n\n') && markToken.slot) {
-				return markToken.slot.end
-			}
-			return markToken.position.end
-		}
-		return (markToken.slot?.start ?? markToken.position.start) + Math.min(offset, nestedLen)
-	}
-
-	let tokenIdx = 0
-	for (const childNode of Array.from(markElement.childNodes)) {
-		if (tokenIdx >= markToken.children.length) break
-		const tokenChild = markToken.children[tokenIdx]
-
-		if (isHtmlElement(childNode) && tokenChild.type === 'text') {
-			if (!isTextTokenSpan(childNode)) continue
-			if (node === childNode) {
-				const charOffset = offset === 0 ? 0 : tokenChild.content.length
-				return tokenChild.position.start + Math.min(charOffset, tokenChild.content.length)
-			}
-			if (childNode.contains(node)) {
-				return tokenChild.position.start + Math.min(offset, tokenChild.content.length)
-			}
-			tokenIdx++
-		} else if (isTextNode(childNode) && tokenChild.type === 'text') {
-			if (node === childNode) {
-				return tokenChild.position.start + Math.min(offset, tokenChild.content.length)
-			}
-			tokenIdx++
-		} else if (isHtmlElement(childNode) && tokenChild.type === 'mark') {
-			if (childNode === node || childNode.contains(node)) {
-				return getDomRawPosInMark(node, offset, childNode, tokenChild)
-			}
-			tokenIdx++
-		}
-	}
-
-	return markToken.slot?.end ?? markToken.position.end
 }
